+++ src/Raven.Studio/typescript/components/common/pathSelector/PathSelector.tsx
@@ -4,7 +4,7 @@ import { Icon } from "components/common/Icon";
 import { LazyLoad } from "components/common/LazyLoad";
 import { LoadError } from "components/common/LoadError";
 import useBoolean from "components/hooks/useBoolean";
-import { useAsyncDebounce } from "components/utils/hooks/useAsyncDebounce";
+import { useAsyncDebounce } from "components/hooks/useAsyncDebounce";
 import React, { useEffect, useImperativeHandle, useState } from "react";
 import { AsyncStateStatus } from "react-async-hook";
 import { Button, Modal, ModalBody, FormGroup, Label, Input, ModalFooter, CloseButton } from "reactstrap";
