+++ src/Raven.Studio/typescript/components/pages/resources/databases/partials/create/formBackup/steps/source/BackupSourceLocal.tsx
@@ -6,7 +6,7 @@ import { Col, Row } from "reactstrap";
 import { CreateDatabaseFromBackupFormData as FormData } from "../../createDatabaseFromBackupValidation";
 import CreateDatabaseFromBackupRestorePoint from "components/pages/resources/databases/partials/create/formBackup/steps/source/RestorePointField";
 import { useRestorePointUtils } from "components/pages/resources/databases/partials/create/formBackup/steps/source/useRestorePointUtils";
-import { useAsyncDebounce } from "components/utils/hooks/useAsyncDebounce";
+import { useAsyncDebounce } from "components/hooks/useAsyncDebounce";
 import EncryptionField from "components/pages/resources/databases/partials/create/formBackup/steps/source/EncryptionField";
 import RestorePointsFields, {
     RestorePointElementProps,
