+++ src/Raven.Studio/typescript/components/pages/resources/databases/partials/create/shared/CreateDatabaseStepDataDirectory.tsx
@@ -5,7 +5,7 @@ import { FormCheckbox, FormPathSelector } from "components/common/Form";
 import { useAppSelector } from "components/store";
 import { clusterSelectors } from "components/common/shell/clusterSlice";
 import { useServices } from "components/hooks/useServices";
-import { useAsyncDebounce } from "components/utils/hooks/useAsyncDebounce";
+import { useAsyncDebounce } from "components/hooks/useAsyncDebounce";
 import { UseAsyncReturn } from "react-async-hook";
 import { CreateDatabaseFromBackupFormData } from "components/pages/resources/databases/partials/create/formBackup/createDatabaseFromBackupValidation";
 import { CreateDatabaseRegularFormData } from "components/pages/resources/databases/partials/create/regular/createDatabaseRegularValidation";
