+++ src/Raven.Studio/typescript/components/pages/resources/databases/partials/create/shared/useCreateDatabaseAsyncValidation.tsx
@@ -2,7 +2,7 @@ import { useServices } from "components/hooks/useServices";
 import { UseFormSetError } from "react-hook-form";
 import { CreateDatabaseRegularFormData } from "../regular/createDatabaseRegularValidation";
 import { CreateDatabaseFromBackupFormData } from "../formBackup/createDatabaseFromBackupValidation";
-import { useAsyncDebounce } from "components/utils/hooks/useAsyncDebounce";
+import { useAsyncDebounce } from "components/hooks/useAsyncDebounce";
 
 export const useCreateDatabaseAsyncValidation = (
     databaseName: string,
